MEDIUM: task: extend the state field to 32 bits
It's been too short for quite a while now and is now full. It's still
time to extend it to 32-bits since we have room for this without
wasting any space, so we now gained 16 new bits for future flags.

The values were not reassigned just in case there would be a few
hidden u16 or short somewhere in which these flags are placed (as
it used to be the case with stream->pending_events).

The patch is tagged MEDIUM because this required to update the task's
process() prototype to use an int instead of a short, that's quite a
bunch of places.

(cherry picked from commit 144f84a09d48e758396ba17f1ddef41b63f30a72)
Add every other prototypes which use a "unsigned short state".
This patch was cherry-picked in 2.3 in order to shut clang warning:
src/mux_h2.c:4085:49: error: implicit conversion from 'int' to 'unsigned short' changes value from -32769 to 32767 [-Werror,-Wconstant-conversion]
        HA_ATOMIC_AND(&h2c->wait_event.tasklet->state, ~TASK_F_USR1);
        ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~^~~~~~~~~~~~~
include/haproxy/atomic.h:270:62: note: expanded from macro 'HA_ATOMIC_AND'
                                     ~~~~~~~~~~~~~~~~~~      ^~~~~

+++ include/haproxy/task.h
@@ -112,7 +112,7 @@ void __task_wakeup(struct task *t, struct eb_root *);
 void __task_queue(struct task *task, struct eb_root *wq);
 
 struct work_list *work_list_create(int nbthread,
-                                   struct task *(*fct)(struct task *, void *, unsigned short),
+                                   struct task *(*fct)(struct task *, void *, unsigned int),
                                    void *arg);
 void work_list_destroy(struct work_list *work, int nbthread);
 unsigned int run_tasks_from_lists(unsigned int budgets[]);
@@ -190,7 +190,7 @@ static inline int thread_has_tasks(void)
  */
 static inline void task_wakeup(struct task *t, unsigned int f)
 {
-	unsigned short state;
+	unsigned int state;
 
 #ifdef USE_THREAD
 	struct eb_root *root;
@@ -347,7 +347,7 @@ static inline struct task *task_unlink_rq(struct task *t)
  */
 static inline void tasklet_wakeup_on(struct tasklet *tl, int thr)
 {
-	unsigned short state = tl->state;
+	unsigned int state = tl->state;
 
 	do {
 		/* do nothing if someone else already added it */
