+++ src/peers.c
@@ -2933,7 +2933,7 @@ static struct appctx *peer_session_create(struct peers *peers, struct peer *peer
  * Task processing function to manage re-connect, peer session
  * tasks wakeup on local update and heartbeat.
  */
-static struct task *process_peer_sync(struct task * task, void *context, unsigned short state)
+static struct task *process_peer_sync(struct task * task, void *context, unsigned int state)
 {
 	struct peers *peers = context;
 	struct peer *ps;
